improve setup tutorial steps
