+++ src/lib/utils/tutorials/nodeActionsTutorialStep.tsx
@@ -10,30 +10,6 @@ export const INITIAL_NODE_TUTORIAL_STATE: TutorialState = null;
 
 dayjs.extend(relativeTime);
 
-/**
-EX: for notebook sections
- "TOOLBAR", "01", "LIVENESS_BAR", "COMMUNITY_LEADERBOARD", "SCROLL_TO_NODE_BUTTON", "FOCUS_MODE_BUTTON"
-Ex for Node id elements to disable
-  "01-close-button",
-  "01-open-button",
-  "01-hide-offsprings-button",
-  "01-hide-button",
-  "01-node-footer-user",
-  "01-node-footer-propose",
-  "01-node-footer-downvotes",
-  "01-node-footer-upvotes",
-  "01-node-footer-tags-citations",
-  "01-button-parent-children",
-  "01-node-footer-ellipsis",
-  "01-reference-button-0"
-  "01-tag-button-0"
-  "01-node-footer-menu"
-  "SearchIcon"
-  "search-recently-input"
-  "recentNodesList"
-  "search-list"
- */
-
 const PARENT_CHILDREN_STEPS: TutorialStepConfig[] = [
   {
     title: "Parent and Children",
@@ -191,10 +167,6 @@ const CLOSE_STEPS: TutorialStepConfig[] = [
   },
 ];
 const EXPAND_STEPS: TutorialStepConfig[] = [];
-
-// export const NODE_CONCEPT_COMPLETE: TutorialStep[] = NODE_CONCEPT_STEPS.map((c, i, s) => {
-//   return { ...getBaseStepConfig(i + 1, s.length), ...c };
-// });
 
 export const PARENT_CHILDREN_STEPS_COMPLETE = PARENT_CHILDREN_STEPS.map((c, i, s) => {
   return { ...getBaseStepConfig(i + 1, s.length), ...c };
