--- src/lib/utils/tutorials/proposalSteps.tsx
@@ -1,300 +1,0 @@
-import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/relativeTime";
-
-import MarkdownRender from "@/components/Markdown/MarkdownRender";
-
-import { TutorialState, TutorialStep, TutorialStepConfig } from "../../../nodeBookTypes";
-// import { FullNodeData, NodeTutorialState, TutorialState } from "../../nodeBookTypes";
-// import { INTERACTIVE_TUTORIAL_NOTEBOOK_NODES } from "../utils/interactiveTutorialNodes";
-import { getBaseStepConfig } from "./tutorial.utils";
-
-export const INITIAL_NODE_TUTORIAL_STATE: TutorialState = null;
-
-dayjs.extend(relativeTime);
-// const STEPS_LENGHT = 47; // 65
-
-/**
-EX: for notebook sections
- "TOOLBAR", "01", "LIVENESS_BAR", "COMMUNITY_LEADERBOARD", "SCROLL_TO_NODE_BUTTON", "FOCUS_MODE_BUTTON"
-Ex for Node id elements to disable
-  "01-close-button",
-  "01-open-button",
-  "01-hide-offsprings-button",
-  "01-hide-button",
-  "01-node-footer-user",
-  "01-node-footer-propose",
-  "01-node-footer-downvotes",
-  "01-node-footer-upvotes",
-  "01-node-footer-tags-citations",
-  "01-button-parent-children",
-  "01-node-footer-ellipsis",
-  "01-reference-button-0"
-  "01-tag-button-0"
-  "01-node-footer-menu"
- */
-
-const OVERVIEW_STEPS: TutorialStepConfig[] = [
-  {
-    title: "Proposals: Overview",
-    description: (
-      <MarkdownRender text={"Most of what you will do on 1Cademy will revolve around making proposals to nodes."} />
-    ),
-  },
-  {
-    title: "Proposals: Overview",
-    description: (
-      <MarkdownRender text={"You will need to propose a change to a node in order to add or edit content."} />
-    ),
-  },
-  {
-    childTargetId: "childNodes",
-    title: "Proposals: Overview",
-    description: (
-      <MarkdownRender
-        text={
-          "You will also have the option to add any of the six types of child nodes which are listed along the right side."
-        }
-      />
-    ),
-  },
-  {
-    title: "Proposals: Overview",
-    description: (
-      <MarkdownRender
-        text={
-          "It is very important to know that new nodes are always added as a proposed child nodes for an existing node, which if accepted, will be the parent node of the one being created."
-        }
-      />
-    ),
-  },
-];
-
-const FINDING_PREREQUISITES_STEPS: TutorialStepConfig[] = [
-  {
-    title: "Proposals: Finding Prerequisites",
-    description: (
-      <MarkdownRender
-        text={
-          "Because new nodes are the children of a node you are editing, an important first step is to determine which node you will be attaching it to."
-        }
-      />
-    ),
-  },
-  {
-    title: "Proposals: Finding Prerequisites",
-    description: (
-      <MarkdownRender
-        text={
-          "Nodes on 1Cademy represent branches of hierarchical relationships where a user will need to know the first concept to understand the concepts linked as children."
-        }
-      />
-    ),
-  },
-  {
-    title: "Proposals: Finding Prerequisites",
-    description: (
-      <MarkdownRender
-        text={
-          "Therefore, you need to know the foundational information before you can understand new information. For example, you need to learn how to count before you can learn to add. "
-        }
-      />
-    ),
-  },
-  {
-    title: "Proposals: Finding Prerequisites",
-    description: (
-      <MarkdownRender
-        text={
-          "When you want to add new content to 1Cademy you need to determine what prerequisite information is required and add your nodes onto the correct prerequisite. If it does not exist then you will need to add it first and link it to the best prerequisite related to what you are adding."
-        }
-      />
-    ),
-  },
-  {
-    childTargetId: "toolbar-search-button",
-    title: "Proposals: Finding Prerequisites",
-    description: (
-      <MarkdownRender
-        text={"In order to see if the prerequisite already exists, you can use the search feature on the toolbar."}
-      />
-    ),
-    tooltipPosition: "right",
-    anchor: "Portal",
-  },
-  {
-    title: "Proposals: Finding Prerequisites",
-    description: (
-      <MarkdownRender
-        text={
-          "You can also manually search through an area of nodes by opening children and parent nodes to find the correct branch of information where you would like to add your nodes."
-        }
-      />
-    ),
-  },
-  {
-    title: "Proposals: Finding Prerequisites",
-    description: (
-      <MarkdownRender
-        text={"Sometimes you will need to build the branch of nodes leading to the information you intend on adding."}
-      />
-    ),
-  },
-];
-const PROPOSING_CONCEPT_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Concept Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit of **Concept Node** you will change the represented **idea** or **concept** of this node, wheter a **superordinate** or **subordinate** Concept"
-        }
-      />
-    ),
-  },
-];
-const PROPOSING_RELATION_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Relation Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Relation Node** will affect multiple concepts identified within the node. It is important to note that the **Relation node** does not define these concepts."
-        }
-      />
-    ),
-  },
-];
-const PROPOSING_REFERENCE_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Reference Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Reference Node** involves adding, modifying, or deleting reference information, Note that purpose is to be cited in other types of nodes."
-        }
-      />
-    ),
-  },
-];
-const PROPOSING_IDEA_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Idea Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Idea Node** involves revising the content of the node to provide **constructive feedback** or suggest new directions for future research. Remember to be concise and clear in your feedback."
-        }
-      />
-    ),
-  },
-];
-const PROPOSING_QUESTION_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Question Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Question Node**  involves modifying the content of a multiple-choice question and its answer options. This can include changing the wording of the question, adding or removing answer options, or adjusting the feedback given to the user after answering the question."
-        }
-      />
-    ),
-  },
-];
-const PROPOSING_CODE_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Code Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Code Node** involves making changes to the displayed code snippet of a specified programming language. This can include modifying existing code, adding new code, or removing existing code using the appropriate programming syntax."
-        }
-      />
-    ),
-  },
-];
-const PROPOSING_EDITS_STEPS: TutorialStepConfig[] = [
-  {
-    childTargetId: "node-body",
-    title: "Proposing Edits - Edit Title",
-    description: <MarkdownRender text={"You can start changing the title"} />,
-  },
-  {
-    childTargetId: "preview-edit",
-    title: "Proposing Edits - Preview or Edit",
-    description: <MarkdownRender text={"You can switch to preview what you have changed before"} />,
-  },
-  {
-    childTargetId: "node-footer-user",
-    title: "Proposing Edits - Contributor",
-    description: (
-      <MarkdownRender
-        text={"Note that the contributor is the person that has done the most to make the node in its present form."}
-      />
-    ),
-  },
-  {
-    childTargetId: "node-type-selector",
-    title: "Proposing Edits - Node Type",
-    description: <MarkdownRender text={"You can also change the type of node here."} />,
-  },
-  {
-    childTargetId: "node-footer-image-video",
-    title: "Proposing Edits - Image or Video",
-    description: <MarkdownRender text={"You can also upload a small image or video"} />,
-  },
-  {
-    childTargetId: "node-footer-tags-citations",
-    title: "Proposing Edits - Tags and Citations",
-    description: (
-      <MarkdownRender
-        text={"You can add and remove **Tags** or **Citations** to references used in the content of the node."}
-      />
-    ),
-  },
-
-  {
-    childTargetId: "button-parent-children",
-    title: "Proposing Edits - Parents and Children",
-    description: <MarkdownRender text={"You can even change parent and children nodes by clicking on this icon."} />,
-  },
-  {
-    childTargetId: "button-propose-proposal",
-    title: "Proposing Edits - Propose",
-    description: (
-      <MarkdownRender
-        text={
-          "Let's assume you just made some changes. Now, it's the time to click this button to submit your proposal. Alternatively, you can click the cancel button, to withdraw your proposal."
-        }
-      />
-    ),
-
-    isClickeable: true,
-  },
-];
-
-console.log(OVERVIEW_STEPS, PROPOSING_EDITS_STEPS, FINDING_PREREQUISITES_STEPS);
-export const PROPOSAL_STEPS_COMPLETE: TutorialStep[] = PROPOSING_EDITS_STEPS.map((c, i, s) => {
-  return {
-    ...getBaseStepConfig(i + 1, s.length),
-    ...c,
-  };
-});
-
-export const PROPOSING_CONCEPT_EDIT_COMPLETE: TutorialStep[] = PROPOSING_CONCEPT_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-export const PROPOSING_RELATION_EDIT_COMPLETE: TutorialStep[] = PROPOSING_RELATION_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-export const PROPOSING_REFERENCE_EDIT_COMPLETE: TutorialStep[] = PROPOSING_REFERENCE_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-export const PROPOSING_IDEA_EDIT_COMPLETE: TutorialStep[] = PROPOSING_IDEA_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-export const PROPOSING_QUESTION_EDIT_COMPLETE: TutorialStep[] = PROPOSING_QUESTION_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-export const PROPOSING_CODE_EDIT_COMPLETE: TutorialStep[] = PROPOSING_CODE_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
