+++ src/lib/utils/tutorials/proposalTutorialSteps.tsx
@@ -1,132 +1,9 @@
-import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/relativeTime";
 import { TutorialState, TutorialStep, TutorialStepConfig } from "src/nodeBookTypes";
 
 import MarkdownRender from "../../../components/Markdown/MarkdownRender";
 import { getBaseStepConfig } from "./tutorial.utils";
 
-// import { FullNodeData, NodeTutorialState, TutorialState } from "../../nodeBookTypes";
-// import { INTERACTIVE_TUTORIAL_NOTEBOOK_NODES } from "../utils/interactiveTutorialNodes";
-
 export const INITIAL_NODE_TUTORIAL_STATE: TutorialState = null;
-
-dayjs.extend(relativeTime);
-// const STEPS_LENGHT = 47; // 65
-
-/**
-EX: for notebook sections
- "TOOLBAR", "01", "LIVENESS_BAR", "COMMUNITY_LEADERBOARD", "SCROLL_TO_NODE_BUTTON", "FOCUS_MODE_BUTTON"
-Ex for Node id elements to disable
-  "01-close-button",
-  "01-open-button",
-  "01-hide-offsprings-button",
-  "01-hide-button",
-  "01-node-footer-user",
-  "01-node-footer-propose",
-  "01-node-footer-downvotes",
-  "01-node-footer-upvotes",
-  "01-node-footer-tags-citations",
-  "01-button-parent-children",
-  "01-node-footer-ellipsis",
-  "01-reference-button-0"
-  "01-tag-button-0"
-  "01-node-footer-menu"
- */
-
-const TMP_EDIT_NODE_CONFIG: TutorialStepConfig[] = [
-  {
-    title: "Propose Node - Edit Node",
-    description: <MarkdownRender text={"**Click** on pencil button to enter in Edit Node"} />,
-    isClickeable: true,
-  },
-];
-
-const TMP_PROPOSE_CHILD_CONCEPT_CONFIG: TutorialStepConfig[] = [
-  {
-    title: "Propose Child - Concept",
-    description: <MarkdownRender text={"**Click** on concept icon button to create a new concept child"} />,
-    isClickeable: true,
-    // childTargetId:""
-  },
-];
-
-const PROPOSING_CONCEPT_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Concept Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit of **Concept Node** you will change the represented **idea** or **concept** of this node, wheter a **superordinate** or **subordinate** Concept"
-        }
-      />
-    ),
-  },
-];
-
-const PROPOSING_RELATION_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Relation Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Relation Node** will affect multiple concepts identified within the node. It is important to note that the **Relation node** does not define these concepts."
-        }
-      />
-    ),
-  },
-];
-
-const PROPOSING_REFERENCE_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Reference Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Reference Node** involves adding, modifying, or deleting reference information, Note that purpose is to be cited in other types of nodes."
-        }
-      />
-    ),
-  },
-];
-
-const PROPOSING_IDEA_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Idea Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Idea Node** involves revising the content of the node to provide **constructive feedback** or suggest new directions for future research. Remember to be concise and clear in your feedback."
-        }
-      />
-    ),
-  },
-];
-
-const PROPOSING_QUESTION_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Question Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Question Node**  involves modifying the content of a multiple-choice question and its answer options. This can include changing the wording of the question, adding or removing answer options, or adjusting the feedback given to the user after answering the question."
-        }
-      />
-    ),
-  },
-];
-
-const PROPOSING_CODE_EDIT: TutorialStepConfig[] = [
-  {
-    title: "Proposing Edits - Code Node",
-    description: (
-      <MarkdownRender
-        text={
-          "Proposing an edit to the **Code Node** involves making changes to the displayed code snippet of a specified programming language. This can include modifying existing code, adding new code, or removing existing code using the appropriate programming syntax."
-        }
-      />
-    ),
-  },
-];
 
 const PROPOSING_EDITS_STEPS: TutorialStepConfig[] = [
   {
@@ -237,41 +114,9 @@ const PROPOSING_EDITS_STEPS: TutorialStepConfig[] = [
   },
 ];
 
-export const TMP_EDIT_NODE: TutorialStep[] = TMP_EDIT_NODE_CONFIG.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-
-export const TMP_PROPOSE_CHILD_CONCEPT: TutorialStep[] = TMP_PROPOSE_CHILD_CONCEPT_CONFIG.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-
 export const PROPOSAL_STEPS_COMPLETE: TutorialStep[] = PROPOSING_EDITS_STEPS.map((c, i, s) => {
   return {
     ...getBaseStepConfig(i + 1, s.length),
     ...c,
   };
-});
-
-export const PROPOSING_CONCEPT_EDIT_COMPLETE: TutorialStep[] = PROPOSING_CONCEPT_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-
-export const PROPOSING_RELATION_EDIT_COMPLETE: TutorialStep[] = PROPOSING_RELATION_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-
-export const PROPOSING_REFERENCE_EDIT_COMPLETE: TutorialStep[] = PROPOSING_REFERENCE_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-
-export const PROPOSING_IDEA_EDIT_COMPLETE: TutorialStep[] = PROPOSING_IDEA_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-
-export const PROPOSING_QUESTION_EDIT_COMPLETE: TutorialStep[] = PROPOSING_QUESTION_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
-});
-
-export const PROPOSING_CODE_EDIT_COMPLETE: TutorialStep[] = PROPOSING_CODE_EDIT.map((c, i, s) => {
-  return { ...getBaseStepConfig(i + 1, s.length), ...c };
 });
